refact: #55 move src/history to src/utils/history refact: #55 move src/store to src/utils/store
