+++ frontend/src/components/QuestionSendFriendItem.test.jsx
@@ -7,7 +7,7 @@ import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 import { act } from 'react-dom/test-utils';
-import history from '../history';
+import history from '../utils/history';
 import rootReducer from '../modules';
 import QuestionSendFriendItem from './QuestionSendFriendItem';
 import { mockStore } from '../mockStore';
