+++ frontend/src/components/QuestionSendModal.test.jsx
@@ -8,7 +8,7 @@ import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 import { act } from 'react-dom/test-utils';
-import history from '../history';
+import history from '../utils/history';
 import rootReducer from '../modules';
 import QuestionSendModal from './QuestionSendModal';
 import { mockStore, mockNoFriendStore } from '../mockStore';
