+++ frontend/src/components/posts/NewPost.test.jsx
@@ -6,7 +6,7 @@ import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 import NewPost from './NewPost';
-import history from '../../history';
+import history from '../../utils/history';
 import rootReducer from '../../modules';
 
 describe('<NewPost />', () => {
