+++ frontend/src/components/posts/QuestionItem.test.jsx
@@ -10,7 +10,7 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 import { act } from 'react-dom/test-utils';
 import axios from '../../utils/api';
-import history from '../../history';
+import history from '../../utils/history';
 import rootReducer from '../../modules';
 import QuestionItem from './QuestionItem';
 import { mockStore } from '../../mockStore';
