+++ frontend/src/components/posts/ShareSettings.test.jsx
@@ -6,7 +6,7 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 import { Router } from 'react-router-dom';
 import { act } from 'react-dom/test-utils';
-import history from '../../history';
+import history from '../../utils/history';
 import rootReducer from '../../modules';
 import ShareSettings from './ShareSettings';
