+++ frontend/src/modules/friend.test.js
@@ -1,6 +1,6 @@
 import axios from '../utils/api';
 import { mockFriendList } from '../constants';
-import store from '../store';
+import store from '../utils/store';
 import friendReducer, * as actionCreators from './friend';
 
 describe('friendActions', () => {
