+++ frontend/src/pages/AnonymousFeed.test.jsx
@@ -9,7 +9,7 @@ import { mockStore } from '../mockStore';
 import rootReducer from '../modules';
 import AnonymousFeed from './AnonymousFeed';
 import 'jest-styled-components';
-import history from '../history';
+import history from '../utils/history';
 
 jest.mock('../components/posts/PostList', () => {
   return jest.fn(() => {
