+++ frontend/src/utils/store/index.js
@@ -1,8 +1,8 @@
-import { createStore, applyMiddleware } from 'redux';
+import { legacy_createStore as createStore, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
-import rootReducer from '../modules';
+import rootReducer from '../../modules';
 
 const store = createStore(
   rootReducer,
